+++ processor/src/test/fixtures/input/com/example/component/PrivateServiceField.java
@@ -1,0 +1,15 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+
+@ArezComponent( allowEmpty = true )
+abstract class PrivateServiceField
+{
+  private final MyService _myService = null;
+
+  @ArezComponent( allowEmpty = true, service = Feature.ENABLE )
+  public abstract static class MyService
+  {
+  }
+}